# schemas/v1/http-api.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "EHAI Execution Plane",
    "version": "2"
  },
  "paths": {
    "/api/v1/notes": {
      "post": {
        "tags": [
          "notes"
        ],
        "summary": "Create Note",
        "operationId": "createNote",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateNoteRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "NoteResponse",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "notes.schema.json#/$defs/NoteResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "notes"
        ],
        "summary": "List Notes",
        "operationId": "listNotes",
        "parameters": [
          {
            "name": "project_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "pattern": "^[0-9A-Fa-f]{8}-[0-9A-Fa-f]{4}-[0-9A-Fa-f]{4}-[0-9A-Fa-f]{4}-[0-9A-Fa-f]{12}$"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Project Id"
            }
          },
          {
            "name": "goal_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "pattern": "^[0-9A-Fa-f]{8}-[0-9A-Fa-f]{4}-[0-9A-Fa-f]{4}-[0-9A-Fa-f]{4}-[0-9A-Fa-f]{12}$"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Goal Id"
            }
          },
          {
            "name": "run_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "pattern": "^[0-9A-Fa-f]{8}-[0-9A-Fa-f]{4}-[0-9A-Fa-f]{4}-[0-9A-Fa-f]{4}-[0-9A-Fa-f]{12}$"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Run Id"
            }
          },
          {
            "name": "include_resolved",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Include Resolved"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "NoteListResponse",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "notes.schema.json#/$defs/NoteListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/notes/{note_id}": {
      "get": {
        "tags": [
          "notes"
        ],
        "summary": "Get Note",
        "operationId": "getNote",
        "parameters": [
          {
            "name": "note_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[0-9A-Fa-f]{8}-[0-9A-Fa-f]{4}-[0-9A-Fa-f]{4}-[0-9A-Fa-f]{4}-[0-9A-Fa-f]{12}$",
              "title": "Note Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "NoteResponse",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "notes.schema.json#/$defs/NoteResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/notes/{note_id}/messages": {
      "post": {
        "tags": [
          "notes"
        ],
        "summary": "Add Note Message",
        "operationId": "addNoteMessage",
        "parameters": [
          {
            "name": "note_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[0-9A-Fa-f]{8}-[0-9A-Fa-f]{4}-[0-9A-Fa-f]{4}-[0-9A-Fa-f]{4}-[0-9A-Fa-f]{12}$",
              "title": "Note Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddNoteMessageRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "NoteResponse",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "notes.schema.json#/$defs/NoteResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/notes/{note_id}/decisions": {
      "post": {
        "tags": [
          "notes"
        ],
        "summary": "Decide Note",
        "operationId": "decideNote",
        "parameters": [
          {
            "name": "note_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[0-9A-Fa-f]{8}-[0-9A-Fa-f]{4}-[0-9A-Fa-f]{4}-[0-9A-Fa-f]{4}-[0-9A-Fa-f]{12}$",
              "title": "Note Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DecideNoteRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "NoteResponse",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "notes.schema.json#/$defs/NoteResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/commands/propose-process": {
      "post": {
        "summary": "Propose Process",
        "operationId": "proposeProcess",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProposeProcessRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Process draft request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "commands.schema.json#/$defs/ProcessDraftAcceptedResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/check-runs/{check_run_id}/decision": {
      "post": {
        "summary": "Decide Human Check",
        "operationId": "decideHumanCheck",
        "parameters": [
          {
            "name": "check_run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[0-9A-Fa-f]{8}-[0-9A-Fa-f]{4}-[0-9A-Fa-f]{4}-[0-9A-Fa-f]{4}-[0-9A-Fa-f]{12}$",
              "title": "Check Run Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DecideHumanCheckRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Run state",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "commands.schema.json#/$defs/RunResponse"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          }
        }
      }
    },
...
    "/api/v1/interventions/{intervention_id}/reply": {
      "post": {
        "summary": "Reply Intervention",
        "operationId": "replyIntervention",
        "parameters": [
          {
            "name": "intervention_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[0-9A-Fa-f]{8}-[0-9A-Fa-f]{4}-[0-9A-Fa-f]{4}-[0-9A-Fa-f]{4}-[0-9A-Fa-f]{12}$",
              "title": "Intervention Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReplyInterventionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Run state",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "commands.schema.json#/$defs/RunResponse"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AddNoteMessageRequest": {
        "properties": {
          "idempotency_key": {
            "type": "string",
            "minLength": 1,
            "title": "Idempotency Key"
          },
          "request_token": {
            "type": "string",
            "minLength": 1,
            "title": "Request Token"
          },
          "actor": {
            "type": "string",
            "minLength": 1,
            "title": "Actor"
          },
          "message": {
            "type": "string",
            "maxLength": 8000,
            "minLength": 1,
            "title": "Message"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "idempotency_key",
          "request_token",
          "actor",
          "message"
        ],
        "title": "AddNoteMessageRequest"
      },
...
      "CreateNoteRequest": {
        "properties": {
          "idempotency_key": {
            "type": "string",
            "minLength": 1,
            "title": "Idempotency Key"
          },
          "goal_id": {
            "type": "string",
            "pattern": "^[0-9A-Fa-f]{8}-[0-9A-Fa-f]{4}-[0-9A-Fa-f]{4}-[0-9A-Fa-f]{4}-[0-9A-Fa-f]{12}$",
            "title": "Goal Id"
          },
          "actor": {
            "type": "string",
            "minLength": 1,
            "title": "Actor"
          },
          "question": {
            "type": "string",
            "maxLength": 8000,
            "minLength": 1,
            "title": "Question"
          },
          "evidence": {
            "type": "string",
            "maxLength": 8000,
            "title": "Evidence",
            "default": ""
          },
          "run_id": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^[0-9A-Fa-f]{8}-[0-9A-Fa-f]{4}-[0-9A-Fa-f]{4}-[0-9A-Fa-f]{4}-[0-9A-Fa-f]{12}$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Run Id"
          },
          "source_kind": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "intervention",
                  "human_check"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Kind"
          },
          "source_id": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^[0-9A-Fa-f]{8}-[0-9A-Fa-f]{4}-[0-9A-Fa-f]{4}-[0-9A-Fa-f]{4}-[0-9A-Fa-f]{12}$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Id"
          },
          "source_token": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Token"
          },
          "origin": {
            "type": "string",
            "enum": [
              "user",
              "agent"
            ],
            "title": "Origin",
            "default": "user"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "idempotency_key",
          "goal_id",
          "actor",
          "question"
        ],
        "title": "CreateNoteRequest"
      },
...
      "DecideHumanCheckRequest": {
        "properties": {
          "idempotency_key": {
            "type": "string",
            "minLength": 1,
            "title": "Idempotency Key"
          },
          "request_token": {
            "type": "string",
            "maxLength": 64,
            "minLength": 64,
            "pattern": "^[0-9a-fA-F]{64}$",
            "title": "Request Token"
          },
          "passed": {
            "type": "boolean",
            "title": "Passed"
          },
          "actor": {
            "type": "string",
            "maxLength": 8000,
            "minLength": 1,
            "title": "Actor"
          },
          "comment": {
            "type": "string",
            "maxLength": 8000,
            "minLength": 1,
            "title": "Comment"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "idempotency_key",
          "request_token",
          "passed",
          "actor",
          "comment"
        ],
        "title": "DecideHumanCheckRequest"
      },
      "DecideNoteRequest": {
        "properties": {
          "idempotency_key": {
            "type": "string",
            "minLength": 1,
            "title": "Idempotency Key"
          },
          "request_token": {
            "type": "string",
            "minLength": 1,
            "title": "Request Token"
          },
          "actor": {
            "type": "string",
            "minLength": 1,
            "title": "Actor"
          },
          "message": {
            "type": "string",
            "maxLength": 8000,
            "minLength": 1,
            "title": "Message"
          },
          "action": {
            "type": "string",
            "enum": [
              "resolve",
              "continue",
              "propose_process",
              "revise_plan"
            ],
            "title": "Action"
          },
          "passed": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Passed"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "idempotency_key",
          "request_token",
          "actor",
          "message",
          "action"
        ],
        "title": "DecideNoteRequest"
      },
...
      "ErrorDetail": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorDetail"
      },
      "ErrorResponse": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        },
        "type": "object",
        "required": [
          "error"
        ],
        "title": "ErrorResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ProposeProcessRequest": {
        "properties": {
          "idempotency_key": {
            "type": "string",
            "minLength": 1,
            "title": "Idempotency Key"
          },
          "run_id": {
            "type": "string",
            "pattern": "^[0-9A-Fa-f]{8}-[0-9A-Fa-f]{4}-[0-9A-Fa-f]{4}-[0-9A-Fa-f]{4}-[0-9A-Fa-f]{12}$",
            "title": "Run Id"
          },
          "reason": {
            "type": "string",
            "maxLength": 16000,
            "minLength": 1,
            "title": "Reason"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "idempotency_key",
          "run_id",
          "reason"
        ],
        "title": "ProposeProcessRequest"
      },
...
      "ReplyInterventionRequest": {
        "properties": {
          "idempotency_key": {
            "type": "string",
            "minLength": 1,
            "title": "Idempotency Key"
          },
          "request_token": {
            "type": "string",
            "maxLength": 64,
            "minLength": 64,
            "pattern": "^[0-9a-fA-F]{64}$",
            "title": "Request Token"
          },
          "actor": {
            "type": "string",
            "maxLength": 8000,
            "minLength": 1,
            "title": "Actor"
          },
          "message": {
            "type": "string",
            "maxLength": 8000,
            "minLength": 1,
            "title": "Message"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "idempotency_key",
          "request_token",
          "actor",
          "message"
        ],
        "title": "ReplyInterventionRequest"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}